+++ src/main/java/de/hs_esslingen/besy/repositories/UserRepository.java
@@ -1,13 +1,14 @@
 package de.hs_esslingen.besy.repositories;
 
-import de.hs_esslingen.besy.models.User;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.Optional;
+import de.hs_esslingen.besy.models.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByKeycloakUUID(String keycloakUUID);
